- extracted the IconsListElement from IconsListWidget into its own file - added JavaDoc
