+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/common/widgets/lists/icons/IconsListWidget.java
@@ -21,35 +21,52 @@ package org.phoenicis.javafx.views.common.widgets.lists.icons;
 import javafx.beans.binding.Bindings;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.geometry.Pos;
 import javafx.scene.CacheHint;
-import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
-import javafx.scene.control.Tooltip;
-import javafx.scene.effect.ColorAdjust;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.FlowPane;
-import javafx.scene.layout.VBox;
-import javafx.scene.shape.Rectangle;
-import org.apache.commons.lang.builder.EqualsBuilder;
-import org.apache.commons.lang.builder.HashCodeBuilder;
 import org.phoenicis.javafx.views.common.MappedList;
 import org.phoenicis.javafx.views.common.widgets.lists.ListWidget;
 import org.phoenicis.javafx.views.common.widgets.lists.ListWidgetEntry;
-import org.phoenicis.javafx.views.common.widgets.lists.icons.StaticMiniature;
 
-import java.util.*;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.function.BiConsumer;
 import java.util.stream.Collectors;
 
+/**
+ * A {@link ListWidget} for a icons list containing a miniature icon and a title for each list entry.
+ *
+ * @author marc
+ * @since 15.05.17
+ */
 public final class IconsListWidget<E> extends ScrollPane implements ListWidget<ListWidgetEntry<E>> {
+    /**
+     * The container of this IconsListWidget that contains the content of this list widget
+     */
     private final FlowPane content;
 
-    private Set<Element<E>> selectedItems;
+    /**
+     * A set of selected items
+     */
+    private Set<IconsListElement<E>> selectedItems;
 
+    /**
+     * A list containing all {@link ListWidgetEntry} objects to be shown inside this {@link ListWidget}
+     */
     private ObservableList<ListWidgetEntry<E>> items;
-    private ObservableList<Element<E>> mappedElements;
 
+    /**
+     * A list containing all mapped {@link IconsListElement} objects to the <code>items</code>
+     */
+    private ObservableList<IconsListElement<E>> mappedElements;
+
+    /**
+     * Constructor
+     *
+     * @param setOnMouseClicked An event listener function to be called when a list element has been selected/clicked
+     */
     public IconsListWidget(BiConsumer<E, MouseEvent> setOnMouseClicked) {
         super();
 
@@ -58,8 +75,8 @@ public final class IconsListWidget<E> extends ScrollPane implements ListWidget<L
         this.selectedItems = new HashSet<>();
 
         this.items = FXCollections.observableArrayList();
-        this.mappedElements = new MappedList<Element<E>, ListWidgetEntry<E>>(items, value -> {
-            Element newElement = new Element(value);
+        this.mappedElements = new MappedList<IconsListElement<E>, ListWidgetEntry<E>>(items, value -> {
+            IconsListElement newElement = new IconsListElement(value);
 
             newElement.setOnMouseClicked(event -> {
                 this.deselectAll();
@@ -100,7 +117,7 @@ public final class IconsListWidget<E> extends ScrollPane implements ListWidget<L
 
     @Override
     public void select(ListWidgetEntry<E> selectedItem) {
-        Element<E> item = this.mappedElements.get(this.items.indexOf(selectedItem));
+        IconsListElement<E> item = this.mappedElements.get(this.items.indexOf(selectedItem));
 
         if (!item.getStyleClass().contains("selected")) {
             item.getStyleClass().add("selected");
@@ -113,77 +130,5 @@ public final class IconsListWidget<E> extends ScrollPane implements ListWidget<L
     public Collection<ListWidgetEntry<E>> getSelectedItems() {
         return this.selectedItems.stream().map(index -> items.get(mappedElements.indexOf(index)))
                 .collect(Collectors.toList());
-    }
-
-    public static class Element<E> extends VBox {
-        private ListWidgetEntry<E> item;
-
-        public Element(ListWidgetEntry<E> item) {
-            super();
-
-            this.item = item;
-
-            this.getStyleClass().add("miniatureListElement");
-            this.setAlignment(Pos.CENTER);
-
-            this.widthProperty().addListener((observable, oldValue, newValue) -> {
-                final Rectangle clip = new Rectangle(this.getWidth(), this.getHeight());
-                this.setClip(clip);
-            });
-
-            this.heightProperty().addListener((observable, oldValue, newValue) -> {
-                final Rectangle clip = new Rectangle(this.getWidth(), this.getHeight());
-                this.setClip(clip);
-            });
-
-            final Label label = new Label(item.getTitle());
-            label.getStyleClass().add("miniatureText");
-
-            StaticMiniature miniature = new StaticMiniature(item.getIconUri());
-
-            this.getChildren().add(miniature);
-            this.getChildren().add(label);
-
-            final Tooltip tooltip = new Tooltip(item.getTitle());
-            Tooltip.install(miniature, tooltip);
-
-            if (!item.isEnabled()) {
-                ColorAdjust grayscale = new ColorAdjust();
-                grayscale.setSaturation(-1);
-                this.setEffect(grayscale);
-            }
-        }
-
-        public E getValue() {
-            return this.item.getItem();
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) {
-                return true;
-            }
-
-            if (o == null || getClass() != o.getClass()) {
-                return false;
-            }
-
-            Element<?> that = (Element<?>) o;
-
-            EqualsBuilder builder = new EqualsBuilder();
-
-            builder.append(item, that.item);
-
-            return builder.isEquals();
-        }
-
-        @Override
-        public int hashCode() {
-            HashCodeBuilder builder = new HashCodeBuilder();
-
-            builder.append(item);
-
-            return builder.toHashCode();
-        }
     }
 }
